requirements.txt
flask
flask_mysqldb
flask_sqlalchemy
flask_restful